src/python/parse_library.py
"""
TestSketcher - Robot Framework Library Parser

This script parses Robot Framework libraries and extracts
keyword information (name, arguments, documentation).
"""

import json
import sys
from robot.api import get_keyword_names, get_keyword_arguments, get_keyword_documentation

def parse_library(library_name):
    """
    Parse a Robot Framework library and extract keyword information.
    
    Args:
        library_name: Name of the library to parse (e.g., 'SeleniumLibrary')
    
    Returns:
        List of keyword dictionaries with name, args, and doc
    """
    # TODO: Implement library parsing
    # This is a placeholder for the actual implementation
    
    keywords = []
    
    # Example structure:
    # keywords.append({
    #     'name': 'Open Browser',
    #     'args': ['url', 'browser=chrome'],
    #     'doc': 'Opens a new browser instance to the given URL.'
    # })
    
    return keywords


def main():
    if len(sys.argv) < 2:
        print("Usage: python parse_library.py <library_name>")
        sys.exit(1)
    
    library_name = sys.argv[1]
    keywords = parse_library(library_name)
    
    print(json.dumps(keywords, indent=2))


if __name__ == '__main__':
    main()
