src/python/requirements.txt
# TestSketcher Python Dependencies

robotframework>=6.0
robotframework-seleniumlibrary>=6.0
